refactor with comments to 6th hw
